Update to v1.6.6
Client elevation now working!

+++ commands/purge.js
@@ -1,4 +1,5 @@
 exports.run = (client, message, args) => {
+  if (!args[0] || args[0] > 100) return message.channel.send(`Please provide a number from 0 to 100 as the first argument.`);
   let messageCount = parseInt(args[0]) + 1;
   message.channel.fetchMessages({limit: messageCount})
     .then(messages => message.channel.bulkDelete(messages));
@@ -8,7 +9,7 @@ exports.conf = {
   enabled: true,
   guildOnly: false,
   aliases: [],
-  permLevel: 2
+  permLevel: 3
 };
 
 exports.help = {
